Update updateDiscounts() method
Discounts should not be over 50 percent, but this was not explicitly defined in the updateDiscounts() method.
I added a bit of logic for that matter.
Note: The test for that part was already in a previous commit.
If it was a real project, I would squash these commits together with an interactive rebase before merging :)